Create folder locker
Windows 10 encrypts folders with a password no third party software is needed. You just need this code above, and make sure that is in a text file. Save the file as .bat open it and it will prompt you for a password  the folder has a name it's locker I think you can change it having messed with this enough to know sorry but it's kind of useful for guest users and hiding personal info from them.  Only down side is if someone was smart enough they would view the .bat file and find the password and gain access to your folder LOL.